requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
httpie==3.2.1
idna==3.3
multidict==6.0.2
platformdirs==2.5.1
pygame==2.1.2
Pygments==2.12.0
PySocks==1.7.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
